Move `SystemdTargetClass` module to real `SystemdTarget` class
