+++ library/systemd/test/test_helper.rb
@@ -3,8 +3,7 @@ require_relative "../../../test/test_helper.rb"
 require "yast2/systemd_unit"
 require "yast2/systemd_service"
 require "yast2/systemd_socket"
-
-Yast.import "SystemdTarget"
+require "yast2/systemd_target"
 
 module SystemctlStubs
   def stub_systemctl(unit)
